+++ test/api_searchProducts_test.dart
@@ -40,7 +40,7 @@ void main() {
 
     test('search favorite products', () async {
       final parameters = <Parameter>[
-        const Page(page: 1),
+        const PageNumber(page: 1),
         const PageSize(size: 10),
         const SortBy(option: SortOption.POPULARITY)
       ];
@@ -66,7 +66,7 @@ void main() {
 
     test('search favorite products EN', () async {
       final parameters = <Parameter>[
-        const Page(page: 14),
+        const PageNumber(page: 14),
         const PageSize(size: 3),
         const SortBy(option: SortOption.EDIT)
       ];
@@ -92,7 +92,7 @@ void main() {
 
     test('type bug : ingredient percent int vs String ', () async {
       final parameters = <Parameter>[
-        const Page(page: 16),
+        const PageNumber(page: 16),
         const PageSize(size: 5),
         const SortBy(option: SortOption.POPULARITY)
       ];
@@ -116,9 +116,9 @@ void main() {
       expect(result.count, greaterThan(30000));
     });
 
-    test('search products by keywords', () async {
+    test('search products by keywords 1', () async {
       final List<Parameter> parameters = <Parameter>[
-        const Page(page: 2),
+        const PageNumber(page: 2),
         const PageSize(size: 10),
         const SearchTerms(terms: ['Kiwi'])
       ];
@@ -140,6 +140,32 @@ void main() {
       expect(result.products!.length, 10);
       expect(result.products![0].runtimeType, Product);
       expect(result.count, greaterThan(900));
+    });
+
+    // Additional test with image field for testing [coordinates_image_size] conversion
+    // c.f. https://github.com/openfoodfacts/openfoodfacts-dart/issues/440
+    test('search products by keywords 2', () async {
+      final List<Parameter> parameters = <Parameter>[
+        const Page(page: 1),
+        const PageSize(size: 10),
+        const SortBy(option: SortOption.POPULARITY),
+        SearchTerms(terms: ['vitamin']),
+      ];
+
+      final ProductSearchQueryConfiguration configuration =
+          ProductSearchQueryConfiguration(
+        parametersList: parameters,
+        language: OpenFoodFactsLanguage.GERMAN,
+        fields: <ProductField>[ProductField.IMAGES],
+      );
+
+      SearchResult result = await OpenFoodAPIClient.searchProducts(
+        null,
+        configuration,
+        queryType: QueryType.PROD,
+      );
+
+      expect(result.products, isNotEmpty);
     });
 
     test('search products filter additives', () async {
@@ -176,7 +202,7 @@ void main() {
 
     test('search products with filter on tags', () async {
       final parameters = <Parameter>[
-        const Page(page: 5),
+        const PageNumber(page: 5),
         const PageSize(size: 10),
         const SortBy(option: SortOption.PRODUCT_NAME),
         TagFilter.fromType(
@@ -384,7 +410,7 @@ void main() {
       );
 
       final parameters = <Parameter>[
-        const Page(page: 1),
+        const PageNumber(page: 1),
         const SearchTerms(terms: ['Quoted Coca "Cola"']),
       ];
 
@@ -486,7 +512,7 @@ void main() {
         fields: [ProductField.ALL],
         parametersList: [
           PnnsGroup2Filter(pnnsGroup2: PnnsGroup2.POTATOES),
-          Page(page: 3),
+          PageNumber(page: 3),
         ],
       );